requirements.txt
boto3
botocore
PyYAML